Consolidate queue worker event listeners
