+++ tests/Unit/Hooks/JobAttemptListenerTest.php
@@ -5,8 +5,6 @@ use Laravel\Nightwatch\Hooks\JobAttemptListener;
 use Tests\FakeJob;
 
 it('gracefully handles exceptions', function () {
-    fakeIngest();
-
     $thrownInJobAttemptSensor = false;
     nightwatch()->sensor->jobAttemptSensor = function () use (&$thrownInJobAttemptSensor) {
         $thrownInJobAttemptSensor = true;
